Share transactions should update company instance
When a company is buying or selling shares of itself then the buy_share
function is given two copies of the company. This is because the get
query returns a new instance every time it is called. So the buy_share
function would retrieve two different versions of the same company,
leading to updates between buy_share and transfer_money to be out of
sync and changes in one function being overwritten by the other.